# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_getUsers",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "path",
            "description": "get user with a specific id",
            "schema": {
              "example": 1234,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "The nymber of entries returned par page",
            "schema": {
              "example": "10",
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "The nymber of entries returned par page",
            "schema": {
              "example": "10",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Fetch user by id",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/posts/{userId}": {
      "get": {
        "operationId": "PostsController_getPosts",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts": {
      "post": {
        "operationId": "PostsController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The post has been created successfully."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Create a new post",
        "tags": [
          "Posts"
        ]
      },
      "patch": {
        "operationId": "PostsController_patchPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/{postId}": {
      "delete": {
        "operationId": "PostsController_deletePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Posts"
        ]
      }
    },
...
    "/tags": {
      "post": {
        "operationId": "TagsController_createTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Tags"
        ]
      }
    },
    "/tags/{tagId}": {
      "delete": {
        "operationId": "TagsController_deleteTag",
        "parameters": [
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Tags"
        ]
      }
    },
    "/tags/{tagId}/soft": {
      "delete": {
        "operationId": "TagsController_softDeleteTag",
        "parameters": [
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Tags"
        ]
      }
    },
...
    "/messages/conversations/{convId}/messages": {
      "post": {
        "operationId": "MessagesController_createMessage",
        "parameters": [
          {
            "name": "convId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message created successfully"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Create message in a conversation",
        "tags": [
          "messages"
        ]
      },
      "get": {
        "operationId": "MessagesController_listMessages",
        "parameters": [
          {
            "name": "convId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List messages in a conversation",
        "tags": [
          "messages"
        ]
      }
    },
    "/messages/conversations/{convId}/read": {
      "post": {
        "operationId": "MessagesController_markRead",
        "parameters": [
          {
            "name": "convId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadConversationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Mark conversation as read up to a seq",
        "tags": [
          "messages"
        ]
      }
    },
    "/messages/conversations/{convId}/unread-count": {
      "get": {
        "operationId": "MessagesController_getUnreadCount",
        "parameters": [
          {
            "name": "convId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get unread message count for conversation",
        "tags": [
          "messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "BlauChat API Documentation",
    "description": "The description of my application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3600/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      },
      "accessToken": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "PostMetaOptionsDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "My First Post",
            "description": "Title of the post"
          },
          "postType": {
            "type": "string",
            "example": "blog",
            "description": "Type of the post"
          },
          "slug": {
            "type": "string",
            "example": "my-first-post",
            "description": "Slug of the post"
          },
          "status": {
            "type": "string",
            "example": "draft",
            "description": "Status of the post"
          },
          "content": {
            "type": "string",
            "example": "This is the content of my first post.",
            "description": "Content of the post"
          },
          "schema": {
            "type": "string",
            "example": "{\"type\":\"object\",\"properties\":{\"body\":{\"type\":\"string\"}}}",
            "description": "JSON schema of the post content"
          },
          "featuredImage": {
            "type": "string",
            "example": "https://example.com/featured-image.jpg",
            "description": "URL of the featured image"
          },
          "publishedOn": {
            "format": "date-time",
            "type": "string",
            "example": "2023-03-15T12:00:00Z",
            "description": "Publication date of the post"
          },
          "tags": {
            "example": [
              "tech",
              "nestjs"
            ],
            "description": "Tags associated with the post",
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "metaOptions": {
            "description": "Key-value meta options for the post",
            "allOf": [
              {
                "$ref": "#/components/schemas/PostMetaOptionsDto"
              }
            ]
          }
        },
        "required": [
          "title",
          "postType",
          "slug",
          "status"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "My First Post",
            "description": "Title of the post"
          },
          "postType": {
            "type": "string",
            "example": "blog",
            "description": "Type of the post"
          },
          "slug": {
            "type": "string",
            "example": "my-first-post",
            "description": "Slug of the post"
          },
          "status": {
            "type": "string",
            "example": "draft",
            "description": "Status of the post"
          },
          "content": {
            "type": "string",
            "example": "This is the content of my first post.",
            "description": "Content of the post"
          },
          "schema": {
            "type": "string",
            "example": "{\"type\":\"object\",\"properties\":{\"body\":{\"type\":\"string\"}}}",
            "description": "JSON schema of the post content"
          },
          "featuredImage": {
            "type": "string",
            "example": "https://example.com/featured-image.jpg",
            "description": "URL of the featured image"
          },
          "publishedOn": {
            "format": "date-time",
            "type": "string",
            "example": "2023-03-15T12:00:00Z",
            "description": "Publication date of the post"
          },
          "tags": {
            "example": [
              "tech",
              "nestjs"
            ],
            "description": "Tags associated with the post",
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "metaOptions": {
            "description": "Key-value meta options for the post",
            "allOf": [
              {
                "$ref": "#/components/schemas/PostMetaOptionsDto"
              }
            ]
          },
          "id": {
            "type": "string",
            "example": "My Updated Post",
            "description": "Id of the post to be updated"
          }
        }
      },
      "CreateTagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Technology",
            "description": "The name of the tag"
          },
          "slug": {
            "type": "string",
            "example": "technology",
            "description": "The slug of the tag"
          },
          "description": {
            "type": "string",
            "example": "A tag for all things technology.",
            "description": "A brief description of the tag"
          },
          "schema": {
            "type": "string"
          },
          "featuredImage": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug",
          "description"
        ]
      },
...
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "clientMsgId": {
            "type": "string",
            "description": "Client generated id for idempotency",
            "example": "cmsg_123"
          },
          "content": {
            "type": "string",
            "description": "Message content",
            "example": "Hello world"
          },
          "attachments": {
            "description": "Attachment urls or ids",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "replyToSeq": {
            "type": "string",
            "description": "Reply to message seq"
          },
          "replyToMessageId": {
            "type": "string",
            "description": "Reply to message id"
          }
        },
        "required": [
          "content"
        ]
      },
      "ReadConversationDto": {
        "type": "object",
        "properties": {
          "seq": {
            "type": "string",
            "description": "Sequence up to which messages are read (inclusive)",
            "example": "150"
          }
        },
        "required": [
          "seq"
        ]
      }
    }
  }
}